chore: set TaskContext and use it in components TaskView shildren components, except TaskDetails
