+++ frontend/src/components/Modal.jsx
@@ -1,15 +1,11 @@
-const Modal = ({ isOpen, onClose, children }) => {
-  if (!isOpen) return null;
+const Modal = ({ onClose, children }) => {
 
   return (
     <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center">
       <div className="bg-gray-800 p-6 rounded-lg shadow-lg max-w-lg w-full">
         <div className="flex justify-end">
-          <button
-            onClick={onClose}
-            className="text-gray-300 text-lg font-bold hover:text-gray-700"
-          >
-            &times;
+          <button onClick={onClose} className="text-gray-300 text-lg font-bold hover:text-gray-700">
+            <span aria-hidden="true">&times;</span>
           </button>
         </div>
         <div className="mt-4">{children}</div>
