+++ frontend/src/components/TaskToolbar.jsx
@@ -1,10 +1,16 @@
+import { useContext } from 'react'
+import { TaskContext } from '../context/TaskContext';
 import TaskForm from "./TaskForm"
 
-export default function TaskToolbar({onTaskAdded}) {
+
+export default function TaskToolbar() {
+
+  const { fetchTasks } = useContext(TaskContext);
+
   return(
     <div className="flex justify-between items-center mb-4">
 
-      <TaskForm onTaskAdded={onTaskAdded} />
+      <TaskForm onTaskAdded={fetchTasks} />
 
     </div>
   );
